nxOMSAutomationWorker ver 1.5.4.0
Linux Hybrid Worker changes for Amazon-Linux
