+++ Providers/nxOMSAutomationWorker/automationworker/worker/automationassets.py
@@ -1,3 +1,7 @@
+#!/usr/bin/env python2
+#
+# Copyright (C) Microsoft Corporation, All rights reserved.
+
 import sys
 
 import configuration
